+++ SayaBot/modules/sql/warns_sql.py
@@ -1,14 +1,14 @@
 import threading
 
 from SayaBot.modules.sql import BASE, SESSION
-from sqlalchemy import Boolean, Column, Integer, String, UnicodeText, distinct, func
+from sqlalchemy import Boolean, Column, Integer, BigInteger, String, UnicodeText, distinct, func
 from sqlalchemy.dialects import postgresql
 
 
 class Warns(BASE):
     __tablename__ = "warns"
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     chat_id = Column(String(14), primary_key=True)
     num_warns = Column(Integer, default=0)
     reasons = Column(postgresql.ARRAY(UnicodeText))
@@ -312,4 +312,5 @@ def migrate_chat(old_chat_id, new_chat_id):
 
 
 __load_chat_warn_filters()
+
